# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ydisks 闲鱼助手 API
  version: 1.0.0
  description: 版本化 HTTP API 的唯一契约源。
servers:
  - url: /
paths:
...
  "/api/v1/automation-issues":
    get:
      operationId: getApiV1AutomationIssues
      summary: getApiV1AutomationIssues
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationIssuesResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/automation-pending-tasks/{task_id}/resolve":
    post:
      operationId: postApiV1AutomationPendingTasksBytask_idResolve
      summary: postApiV1AutomationPendingTasksBytask_idResolve
      parameters:
          - name: task_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution]
              properties:
                resolution:
                  type: string
                  enum: [retry, dismiss, stop_order]
                  description: stop_order 持久停止该异常关联订单的全部自动化，不撤销已发送或在途请求；缺少可靠订单身份时拒绝。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/v1/automation-rules":
    get:
      operationId: getApiV1AutomationRules
      summary: getApiV1AutomationRules
      security: [{cookieAuth: []}]
      parameters:
        - name: page
          in: query
          required: false
          schema: {type: integer, minimum: 1}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100}
        - name: cookie_id
          in: query
          required: false
          schema: {type: string}
        - name: trigger_type
          in: query
          required: false
          schema: {type: string}
        - name: enabled
          in: query
          required: false
          schema: {type: boolean}
        - name: search
          in: query
          required: false
          schema: {type: string}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationRulesResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: postApiV1AutomationRules
      summary: postApiV1AutomationRules
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationRuleRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 配置冲突或发货模板状态已变化
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/automation-rules/{rule_id}":
    delete:
      operationId: deleteApiV1AutomationRulesByrule_id
      summary: deleteApiV1AutomationRulesByrule_id
      parameters:
          - name: rule_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: putApiV1AutomationRulesByrule_id
      summary: putApiV1AutomationRulesByrule_id
      parameters:
          - name: rule_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationRuleRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 配置冲突或发货模板状态已变化
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/automation-runs/{run_id}/resolve":
    post:
      operationId: postApiV1AutomationRunsByrun_idResolve
      summary: postApiV1AutomationRunsByrun_idResolve
      parameters:
          - name: run_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution]
              properties:
                resolution:
                  type: string
                  enum: [continue, retry, cancel, stop_order]
                  description: stop_order 持久停止该异常关联订单的全部自动化，不撤销已发送或在途请求；缺少可靠订单身份时拒绝。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
      description: 统一响应的稳定字段；业务成功字段由各阶段 operation schema 收紧。
...
    AutomationRunIssue:
      type: object
      required: [id, cookie_id, order_id, trigger_type, error_message, issue_kind, allowed_resolutions, action_cursor, sent_count, updated_at]
      properties:
        account_name: {type: string}
        item_id: {type: string}
        item_title: {type: string}
        buyer_id: {type: string}
        chat_id: {type: string}
        order_status: {type: string}
        can_stop_order: {type: boolean}
        id: {type: integer}
        cookie_id: {type: string}
        order_id: {type: string}
        trigger_type: {type: string}
        error_message: {type: string}
        issue_kind: {type: string}
        allowed_resolutions: {type: array, items: {type: string}}
        action_cursor: {type: integer}
        sent_count: {type: integer}
        updated_at: {type: string}
    DeferredAutomationIssue:
      type: object
      required: [id, cookie_id, trigger_type, error_message, attempt_count, updated_at]
      properties:
        account_name: {type: string}
        item_id: {type: string}
        item_title: {type: string}
        buyer_id: {type: string}
        chat_id: {type: string}
        order_status: {type: string}
        can_stop_order: {type: boolean}
        order_id: {type: string}
        id: {type: integer}
        cookie_id: {type: string}
        trigger_type: {type: string}
        error_message: {type: string}
        attempt_count: {type: integer}
        updated_at: {type: string}
    AutomationIssuesResponse:
      type: object
      required: [runs, pending_tasks]
      properties:
        runs: {type: array, items: {$ref: '#/components/schemas/AutomationRunIssue'}}
        pending_tasks: {type: array, items: {$ref: '#/components/schemas/DeferredAutomationIssue'}}
    AutomationTemplateBinding:
      type: object
      required: [variable_key, card_id, delivery_count]
      properties:
        variable_key: {type: string}
        card_id: {type: integer}
        delivery_count: {type: integer}
    AutomationTemplateBindingRequest:
      type: object
      required: [key, card_id, delivery_count]
      properties:
        key: {type: string}
        card_id: {type: integer}
        delivery_count: {type: integer}
    AutomationActionRequest:
      type: object
      required: [action_type, card_id, delivery_count, message_template, delay_seconds, config_json, enabled, sort_order, delivery_template_id, template_bindings, custom_variables]
      properties:
        id: {type: integer}
        action_type: {type: string}
        card_id: {type: integer}
        delivery_count: {type: integer}
        message_template: {type: string}
        delay_seconds: {type: integer}
        config_json: {type: string}
        enabled: {type: boolean}
        sort_order: {type: integer}
        delivery_template_id: {type: integer}
        template_bindings: {type: array, items: {$ref: '#/components/schemas/AutomationTemplateBindingRequest'}}
        custom_variables: {type: object, additionalProperties: {type: string}}
    AutomationRuleRequest:
      type: object
      required: [cookie_id, item_id, name, trigger_type, enabled, priority, config_json, actions]
      properties:
        cookie_id: {type: string}
        item_id: {type: string}
        name: {type: string}
        trigger_type: {type: string}
        enabled: {type: boolean}
        priority: {type: integer}
        config_json: {type: string, description: '规则 JSON 配置；账号级 order_paid 仅在 allow_all_items 为布尔 true 时允许全商品兜底，省略或 false 均不授权。'}
        actions: {type: array, items: {$ref: '#/components/schemas/AutomationActionRequest'}}
    AutomationActionResponse:
      type: object
      required: [id, action_type, card_id, card_name, delivery_count, message_template, delay_seconds, config_json, enabled, sort_order, delivery_template_id, delivery_template_name, template_keys, template_bindings, custom_variables]
      properties:
        id: {type: integer}
        action_type: {type: string}
        card_id: {type: integer}
        card_name: {type: string}
        delivery_count: {type: integer}
        message_template: {type: string}
        delay_seconds: {type: integer}
        config_json: {type: string}
        enabled: {type: boolean}
        sort_order: {type: integer}
        delivery_template_id: {type: integer}
        delivery_template_name: {type: string}
        template_keys: {type: array, items: {type: string}}
        template_bindings: {type: array, items: {$ref: '#/components/schemas/AutomationTemplateBinding'}}
        custom_variables: {type: object, additionalProperties: {type: string}}
    AutomationRuleResponse:
      type: object
      required: [id, cookie_id, item_id, item_title, name, trigger_type, enabled, priority, config_json, sku_migration_status, actions, created_at, updated_at]
      properties:
        id: {type: integer}
        cookie_id: {type: string}
        item_id: {type: string}
        item_title: {type: string}
        name: {type: string}
        trigger_type: {type: string}
        enabled: {type: boolean}
        priority: {type: integer}
        config_json: {type: string, description: '规则 JSON 配置；账号级 order_paid 仅在 allow_all_items 为布尔 true 时允许全商品兜底，省略或 false 均不授权。'}
        sku_migration_status: {type: string, enum: [pending, ready, needs_reconfiguration]}
        actions: {type: array, items: {$ref: '#/components/schemas/AutomationActionResponse'}}
        created_at: {type: string}
        updated_at: {type: string}
    AutomationRuleListResponse:
      type: array
      items: {$ref: '#/components/schemas/AutomationRuleResponse'}
    AutomationRulePageResponse:
      type: object
      required: [success, data, total, page, page_size, total_pages, trigger_counts]
      properties:
        success: {type: boolean}
        data: {type: array, items: {$ref: '#/components/schemas/AutomationRuleResponse'}}
        total: {type: integer}
        page: {type: integer}
        page_size: {type: integer}
        total_pages: {type: integer}
        trigger_counts:
          type: object
          additionalProperties: {type: integer}
    AutomationRulesResponse:
      oneOf:
        - $ref: '#/components/schemas/AutomationRuleListResponse'
        - $ref: '#/components/schemas/AutomationRulePageResponse'
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
      additionalProperties: true